requirements.txt
openai>=1.42.0
python-dotenv>=1.0.1
tenacity>=9.0.0
pydantic>=2.8.2
